Refactor component readiness query generation
Stop passing the bulk of the query into the query generators and let
them generate it themselves internally. Introduce constructor methods to
help cut down on instantiation of query generators without some param we
expect to be set.
